Create repos

+++ src/main/java/com/example/shopsphere/repository/CategoriesRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.shopsphere.entity.Categories;
 
-public interface CategoriesRepository extends JpaRepository<Categories, Integer> {
+public interface CategoriesRepository extends JpaRepository<Categories, Long> {
 
     
 }
