+++ src/main/java/com/example/shopsphere/repository/ReviewsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.shopsphere.entity.Reviews;
 
-public interface ReviewsRepository extends JpaRepository<Reviews, Integer>{
+public interface ReviewsRepository extends JpaRepository<Reviews, Long>{
     
 }
